move priceFormatter
