+++ app/helpers/contextCreators/priceFormatter.js
