+++ app/pages/preview/contextCreator.js
@@ -1,7 +1,7 @@
 import manifest from './manifest.json';
 import { baseUrl } from '../../config';
 import { formatDate } from '../../helpers/dateFormatter';
-import { formatPrice } from '../../helpers/priceFormatter';
+import { formatPrice } from '../../helpers/contextCreators/priceFormatter';
 import { logger } from '../../logger';
 
 const generateCallOffPartyDetails = ({ orderPartyData }) => {
